+++ source/entities/cell/cell.cpp
@@ -14,7 +14,7 @@ Cell::Cell( sf::Vector2f position, sf::Vector2f cellDimensions, sf::Texture& tex
 
 	this->getComponent<ProperBody>()->getAs<sf::Sprite>().setPosition( position );
 	this->getComponent<ProperBody>()->getAs<sf::Sprite>().setTexture( texture );
-	this->getComponent<ProperBody>()->getAs<sf::Sprite>().setTextureRect( sf::IntRect( id*cellDimensions.x, 0, cellDimensions.x, cellDimensions.y ) );
+	this->getComponent<ProperBody>()->getAs<sf::Sprite>().setTextureRect( sf::IntRect( static_cast<int>(id*cellDimensions.x), 0, static_cast<int>(cellDimensions.x), static_cast<int>(cellDimensions.y) ) );
 
 	this->getComponent<UnitPosition>()->setUnitPosition( Math::convertToUnitPosition( position, cellDimensions ) );
 }
